fix delete and finish e2e tests

+++ sync/src/main/java/io/apicurio/sync/controller/ArtifactController.java
@@ -12,11 +12,13 @@ import org.slf4j.LoggerFactory;
 import io.apicurio.registry.rest.client.RegistryClient;
 import io.apicurio.registry.rest.client.exception.ArtifactAlreadyExistsException;
 import io.apicurio.registry.rest.client.exception.ArtifactNotFoundException;
+import io.apicurio.registry.rest.client.exception.VersionNotFoundException;
 import io.apicurio.registry.rest.v2.beans.ArtifactMetaData;
 import io.apicurio.registry.rest.v2.beans.EditableMetaData;
 import io.apicurio.registry.rest.v2.beans.IfExists;
 import io.apicurio.registry.rest.v2.beans.UpdateState;
 import io.apicurio.registry.rest.v2.beans.VersionMetaData;
+import io.apicurio.registry.rest.v2.beans.VersionSearchResults;
 import io.apicurio.registry.types.ArtifactState;
 import io.apicurio.registry.types.ArtifactType;
 import io.apicurio.sync.Configuration;
@@ -57,13 +59,38 @@ public class ArtifactController implements ResourceController<Artifact> {
         if (config.getDeleteArtifactsEnabled()) {
             log.debug("Handling delete {}", resource.getMetadata().getName());
 
-            ArtifactSpec spec = resource.getSpec();
+            if (resource.getMetadata().getDeletionTimestamp() != null) {
+                log.debug("del - Deletion timestamp is not null, this may be a marked for deletion object");
+            }
+
+            ArtifactStatus spec = resource.getStatus();
 
             var artifactsResources = artifactResourceClient.listVersions(resource);
             if (artifactsResources.getItems().isEmpty()) {
-                registryClient.deleteArtifact(spec.getGroupId(), spec.getArtifactId());
+                log.debug("Going to delete artifact {} {}", spec.getGroupId(), spec.getArtifactId());
+                try {
+                    registryClient.deleteArtifact(spec.getGroupId(), spec.getArtifactId());
+                } catch (ArtifactNotFoundException e) {
+                    //ignored
+                }
             } else {
-                log.debug("Not deleteing, resources {}", artifactsResources.getItems().toString());
+                log.debug("Not deleteing, resources found {}", artifactsResources.getItems().size());
+                try {
+                    UpdateState update = new UpdateState();
+                    update.setState(ArtifactState.DISABLED);
+                    registryClient.updateArtifactVersionState(spec.getGroupId(), spec.getArtifactId(), spec.getVersion(), update);
+                } catch (VersionNotFoundException e) {
+                    //ignored
+                    return DeleteControl.DEFAULT_DELETE;
+                }
+
+                VersionSearchResults versions = registryClient.listArtifactVersions(spec.getGroupId(), spec.getArtifactId(), 0, 100);
+                boolean notAllDisabled = versions.getVersions().stream()
+                    .anyMatch(ver -> ver.getState() != ArtifactState.DISABLED);
+                if (!notAllDisabled) {
+                    log.debug("Going to delete artifact, all versions disabled {} {}", spec.getGroupId(), spec.getArtifactId());
+                    registryClient.deleteArtifact(spec.getGroupId(), spec.getArtifactId());
+                }
             }
 
         }
@@ -72,6 +99,10 @@ public class ArtifactController implements ResourceController<Artifact> {
 
     @Override
     public UpdateControl<Artifact> createOrUpdateResource(Artifact resource, Context<Artifact> context) {
+
+        if (resource.getMetadata().getDeletionTimestamp() != null) {
+            log.debug("Deletion timestamp is not null, this may be a marked for deletion object");
+        }
 
         log.debug("Handling createOrUpdate {}", resource.getMetadata().getName());
